Format All Variables to Use American Standard for "Colour"
For less confusion in the codebase for both myself, and anyone who might
work on this in the future, standardise all uses of "colour" in code and
replace with American spelling, "color". This way, it confirms to code
language standards despite my Australian education screaming at me in
protest to the absurdity of American English.